+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Added functionality to use EPSG codes and Extents in the query parameters
 - Fixed a Bug where a long AppTitle would not be shown correctly
 - Removes `mapVersion` from plugin serialization. The `mapVersion` was and is set by the app regardless of the plugin's config.
+- Fix `VcsList` issue, when clicking list items of a not selectable list. Remove deprecated unused api `add`, `remove` and `clear` from `VcsList`.
 
 # 6.0.9